requirements.txt
fastapi
uvicorn
influxdb
python-dotenv
playwright>=1.43
garminconnect 
psycopg2-binary 
python-dateutil
requests